PDBs_1cl.txt PDBs_big_cl.txt
Lists of pdb codes that are representative for each cluster. Still have to summarize the ~500 clusters that have more than one structure but low similarity. 
In fact, biologically probably those are the most interesting: the proteins that have been more times studied with such different ligands that its conformation changes a lot.